helm/ai-agent-connector/tests/requirements.txt
# Test dependencies for Helm chart tests
pytest>=7.0.0
pyyaml>=6.0
